Move XML helpers and inline project fragments
