+++ src/MooVC.Syntax.CSharp/Attributes/Project/Import.cs
@@ -3,6 +3,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.Linq;
+    using System.Xml.Linq;
     using Fluentify;
     using Valuify;
     using Ignore = Valuify.IgnoreAttribute;
@@ -42,6 +43,26 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
                 .And(nameof(Project), _ => !Project.IsSingleLine, Project)
                 .And(nameof(Sdk), _ => !Sdk.IsMultiLine, Sdk)
                 .Results;
+        }
+
+        public XElement ToFragment()
+        {
+            return new XElement(
+                nameof(Import),
+                Project.ToXmlAttribute(nameof(Project)),
+                Sdk.ToXmlAttribute(nameof(Sdk)),
+                Condition.ToXmlAttribute(nameof(Condition)),
+                Label.ToXmlAttribute(nameof(Label)));
+        }
+
+        public override string ToString()
+        {
+            if (IsUndefined)
+            {
+                return string.Empty;
+            }
+
+            return ToFragment().ToString();
         }
     }
 }
