+++ src/MooVC.Syntax.CSharp/Attributes/Project/ItemGroup.cs
@@ -4,6 +4,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
     using System.Collections.Immutable;
     using System.ComponentModel.DataAnnotations;
     using System.Linq;
+    using System.Xml.Linq;
     using Fluentify;
     using Valuify;
     using Ignore = Valuify.IgnoreAttribute;
@@ -40,6 +41,29 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
                 .AndIf(!Items.IsDefaultOrEmpty, nameof(Items), item => !item.IsUndefined, Items)
                 .And(nameof(Label), _ => !Label.IsMultiLine, Label)
                 .Results;
+        }
+
+        public XElement ToFragment()
+        {
+            var items = Items
+                .Where(item => !item.IsUndefined)
+                .Select(item => item.ToFragment());
+
+            return new XElement(
+                nameof(ItemGroup),
+                Condition.ToXmlAttribute(nameof(Condition)),
+                Label.ToXmlAttribute(nameof(Label)),
+                items);
+        }
+
+        public override string ToString()
+        {
+            if (IsUndefined)
+            {
+                return string.Empty;
+            }
+
+            return ToFragment().ToString();
         }
     }
 }
