--- src/MooVC.Syntax.CSharp/Attributes/Project/TaskParameter.ToFragment.cs
@@ -1,25 +1,0 @@
-namespace MooVC.Syntax.CSharp.Attributes.Project
-{
-    using System.Xml.Linq;
-
-    public sealed partial class TaskParameter
-    {
-        public XElement ToFragment()
-        {
-            return new XElement(
-                "Parameter",
-                ProjectXml.Attribute(nameof(Name), Name),
-                Value.IsEmpty ? null : Value.ToString());
-        }
-
-        public override string ToString()
-        {
-            if (IsUndefined)
-            {
-                return string.Empty;
-            }
-
-            return ToFragment().ToString();
-        }
-    }
-}
