+++ src/MooVC.Syntax.CSharp/Concepts/Project.cs
@@ -4,6 +4,8 @@
     using System.Collections.Generic;
     using System.Collections.Immutable;
     using System.ComponentModel.DataAnnotations;
+    using System.Linq;
+    using System.Xml.Linq;
     using Fluentify;
     using MooVC.Syntax.CSharp.Attributes.Project;
     using Valuify;
@@ -47,6 +49,54 @@
                 .AndIf(!Sdks.IsDefaultOrEmpty, nameof(Sdks), sdk => !sdk.IsUnspecified, Sdks)
                 .AndIf(!Targets.IsDefaultOrEmpty, nameof(Targets), target => !target.IsUndefined, Targets)
                 .Results;
+        }
+
+        public XDocument ToDocument()
+        {
+            if (IsUndefined)
+            {
+                return new XDocument();
+            }
+
+            var imports = Imports
+                .Where(import => !import.IsUndefined)
+                .Select(import => import.ToFragment());
+
+            var itemGroups = ItemGroups
+                .Where(group => !group.IsUndefined)
+                .Select(group => group.ToFragment());
+
+            var propertyGroups = PropertyGroups
+                .Where(group => !group.IsUndefined)
+                .Select(group => group.ToFragment());
+
+            var sdks = Sdks
+                .Where(sdk => !sdk.IsUnspecified)
+                .Select(sdk => sdk.ToFragment());
+
+            var targets = Targets
+                .Where(target => !target.IsUndefined)
+                .Select(target => target.ToFragment());
+
+            var project = new XElement(
+                nameof(Project),
+                imports,
+                itemGroups,
+                propertyGroups,
+                sdks,
+                targets);
+
+            return new XDocument(project);
+        }
+
+        public override string ToString()
+        {
+            if (IsUndefined)
+            {
+                return string.Empty;
+            }
+
+            return ToDocument().ToString();
         }
     }
 }
