+++ src/MooVC.Syntax.CSharp/XmlExtensions.cs
@@ -1,13 +1,14 @@
-namespace MooVC.Syntax.CSharp.Attributes.Project
+namespace MooVC.Syntax.CSharp
 {
     using System.Xml.Linq;
+    using MooVC.Syntax.CSharp.Attributes.Project;
     using MooVC.Syntax.CSharp.Elements;
 
-    internal static class ProjectXml
+    internal static class XmlExtensions
     {
         private static readonly Identifier.Options IdentifierOptions = Identifier.Options.Pascal;
 
-        internal static XAttribute Attribute(string name, Snippet value)
+        internal static XAttribute ToXmlAttribute(this Snippet value, string name)
         {
             if (value.IsEmpty)
             {
@@ -17,7 +18,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
             return new XAttribute(name, value.ToString());
         }
 
-        internal static XAttribute Attribute(string name, Identifier value)
+        internal static XAttribute ToXmlAttribute(this Identifier value, string name)
         {
             if (value.IsUnnamed)
             {
@@ -27,7 +28,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
             return new XAttribute(name, value.ToSnippet(IdentifierOptions).ToString());
         }
 
-        internal static XAttribute Attribute(string name, Qualifier value)
+        internal static XAttribute ToXmlAttribute(this Qualifier value, string name)
         {
             if (value.IsUnqualified)
             {
@@ -37,7 +38,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
             return new XAttribute(name, value.ToString());
         }
 
-        internal static XAttribute Attribute(string name, bool value)
+        internal static XAttribute ToXmlAttribute(this bool value, string name)
         {
             if (!value)
             {
@@ -47,7 +48,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
             return new XAttribute(name, value.ToString().ToLowerInvariant());
         }
 
-        internal static XAttribute Attribute(string name, TargetTask.Options value)
+        internal static XAttribute ToXmlAttribute(this TargetTask.Options value, string name)
         {
             if (value == TargetTask.Options.ErrorAndStop)
             {
@@ -57,7 +58,7 @@ namespace MooVC.Syntax.CSharp.Attributes.Project
             return new XAttribute(name, value.ToString());
         }
 
-        internal static string ElementName(Identifier value, string fallback)
+        internal static string ToXmlElementName(this Identifier value, string fallback)
         {
             if (value.IsUnnamed)
             {
